Order detail profile working fine, adding product to cart and saving it working fine, updating the amount in cart working fine

+++ CapstoneProject/src/main/java/com/qvainside/CapstoneProject/database/entity/OrderDetail.java
@@ -34,7 +34,7 @@ public class OrderDetail {
     private Date bookingDate;
 
     @Column(name = "duration_hours", nullable = false, precision = 10)
-    private Double durationHours;
+    private Integer durationHours;
 
     @Column(name = "number_of_pax", nullable = false)
     private Integer numberOfPax;
